2014-12-01. (A). Added a "prompt" text to the two combo boxes in Food tab to instruct users to first select a building and then a process. (B). Capitalize resource and part names in Resupply Tool. (C). Misc. fixes
Details:
(1). Previously the combo box shows an empty selection on the buildingSelection and processSelection list.
e.g.
buildingSelection.setRenderer(new PromptComboBoxRenderer(" Select a Building"));
buildingSelection.setSelectedIndex(-1);

(2). Added <resource type="carbon dioxide" amount="1000.0" />
to Mars Direct Base (phase 1) in settlements.xml 
(Question: is it supposed to be zero as intended? )

(3). Fixed the extraneous tooltip text descriptions for "Override Food Production" in TabPanelBuildings.java and "Override Manufacturing" in TabPanelManufacture.java

(4). Capitalize first letter of each item of the Supplies List with 
WordUtils.capitalize() in the Resupply Tool:  
a. Resource names in createResourcesDisplayPanel() and part names in createPartsDisplayPanel() in ResupplyDetailPanel.java. 
b. Names of "type" in TypeCellEditor() of in ResupplyMissionEditingPanel.java.  
c. Resource and part names in SupplyTableModel.java. 

(5). Avoided differences of nomenclature by eliminating the word "Group" in each crop in crops.xml and through all classes.
e.g. If a crop belongs to the Vegetable Group, its crop-category is "Vegetables".