2015-01-09 (A). Improved the demand system for the amount resources by adding 2 addition demand metrics.  The new features tracked the # of demand successful and unsuccessful requests and the demand quantity and were integrated into the existing demand/supply system.
(1). Added a number of formulae to calculate the total average yearly demand in determineAmountResourceGoodValue() in GoodsManager.java;  

(2). Added a number of methods and Maps in Inventory.java;  

(3). Call one of the two methods when a demand is made on an amount resource in FoodProduction.java, PhysicalCondition.java, Cooking.java, eatMeal.java, EVASuit.java, MalfunctionManager.java, BuildingConstructionMission.java, Settlement.java, Crop.java, LivingAccommodations.java, and ResourceProcess.java, Rover.java, FoodProductionUtil.java,
BuildingConstructionMission.java, EmergencySupplyMission.java, ManufactureUtil.java, MalfunctionManager.java, etc. :

a. Add a new method addDemandRealUsage() when retrieveAmountResource() is called. 
It's for tracking the demand quantity made in a successful request . 

b. Add a new method addDemandTotalRequest() when  getAmountResourceStored() is called. 
It's for calculating the # of times a demand was made in a request (the request is either successful and unsuccessful).  

NOTE: if the amount resource is NOT available, then it is considered an unsuccessful request. 

(4). Added a sample daily demand statistics into a new method called calculateDemand() in Settlement.java.