2015-01-15 (A). Implemented person detection for the right-click popup menu. One can now right click on a person and choose menu item "Detail" to fire up the person's unit window DIRECTLY in Settlement Map Tool);
(B). Fixed description panel when a vehicle is detected in right click popup menu;  

(C). Set up proper if-else hierarchy in doPop() that deconflicts cases when a person and/ or building and/or vehicle are detected (say, if a person is within a vehicle which is within a building or if a vehicle is within a building) in SettlementWindow.java.

(D). Corrected the case when both showMedical and showMalfunction are false, no notification messages should be fired in eventAdded() in EventTableModel.java. 

(E). Added the orange color background to the column headers in the Monitor Tool's Historical Event Tab. Changed codes in TableTab.java to load TableHeaderRenderer class in both ordered and unordered table model.