2015-01-16a (A). Switched to using Msg.getString("temperature.sign.degreeCelsius") in LocationTabPanel.java for the degree sign instead of using " Â°C".
Question with a concern: has it been implemented this way in the past? Will this change pass the JUnit test?

(B). Reverted the value of the parameter WASTE_WATER_NEEDED back to 5D from 2D in Crop.java.

(C). Deleted the line that divided projectedDemand by 667. This affects how things are triggered in msp such as water recycling process in Lander Hab. Will tune up these threshold that trigger various processes in future.