2015-01-17 (A). Fixed the incorrect building nickname assignment for buildings in the initial settlement and from the standard resupply missions or subsequent manual resupply mission. Each settlement has a scenario id (determined by the settlement template being used) that will be translated into a alphabet. e.g.  0 ==> A, 1 ==> B, 2 ==> C, etc..
Note that building nickname is not unique in a simulation. For example, if a user has two Mars Direct Phase 1 settlements in a simulation, then there will be two "Inflatable Greenhouse A1", one in each settlement.

In future, we can make each settlement having an unique id, instead of using scenario id.

(B).  Relocated the pause announcement window (& the Transport Wizard JOption box) to  3/4 from the top and 3/4 from the left so that it would NOT block the view of the settlement name comboBox at the top center.