2015-01-19a Added setDefaultCloseOperation(WindowConstants.HIDE_ON_CLOSE) in SettlementWindow.java so that when a user clicks the close button, the toolWindow will become hidden or invisible, instead of being disposed. The cost is the memory is not released and not being garbage collected.
Note 1 : This is only a TEMPORARY workaround to shorten the delay time.

Note 2 : before this patch, it always takes two seconds on my core i7-4700MQ laptop and probably up to five seconds on a slow machine to re-create the Settlement Map Tool.

Note 3: SettlementWindow extends ToolWindow and ToolWindow extends JInternalFrame

Note 4: I don't mind dispose all other unused toolWindow. But the Settlement Map Tool should probably NOT be disposed because of its slow rendering speed.

Note 5: In ToolFrameListener.java, we have:

	public void internalFrameOpened(InternalFrameEvent e) {
		JInternalFrame frame = (JInternalFrame) e.getSource();
		try { frame.setClosed(false); }
		......
		
So, if a user re-open a toolWindow, it will try to setClosed(false) first---to make an invisible toolWindow visible, before it tries to re-create a toolWindow from scratch.		
		
Other observations: 

a). On both 3.06 and 3.07 (the release version), when a simulation first gets started, I always have to hit function key TWICE (F7 on 3.06, F6 on 3.07) to bring up the Settlement Map Tool. 

b). It's also true after I hit the closed button to dispose the Settlement Map Tool, if I want to see it again, I ALWAYS have to hit the function key TWICE. Why? 

c). I'm guessing that: could it be true that, when first hitting F6, the settlement map tool is created, but is not visible? It's NOT until hitting F6 again that the tool becomes visible !?!?

Question: can someone fix this in openToolWindow() in MainDesktopPane.java?


In MainWindowMenu.java,

		if (selectedItem == settlementToolItem) {
			if (settlementToolItem.isSelected()) desktop.openToolWindow(SettlementWindow.NAME);
			else desktop.closeToolWindow(SettlementWindow.NAME);
		}

In MainDesktopPane.java,
		
	public void closeToolWindow(String toolName) {
		ToolWindow window = getToolWindow(toolName);
		if ((window != null) && !window.isClosed()) {
			try { window.setClosed(true); }
			catch (java.beans.PropertyVetoException e) {}
		}
	}		
	
	public void openToolWindow(String toolName) {
		ToolWindow window = getToolWindow(toolName);
		if (window != null) {
			if (window.isClosed()) {
				if (!window.wasOpened()) {
					UIConfig config = UIConfig.INSTANCE;
					if (config.useUIDefault()) {
						window.setLocation(getCenterLocation(window));
					} else {
						if (config.isInternalWindowConfigured(toolName)) {
							window.setLocation(config.getInternalWindowLocation(toolName));
							if (window.isResizable()) {
								window.setSize(config.getInternalWindowDimension(toolName));
							}
						} else {
							window.setLocation(getRandomLocation(window));
						}
					}
					window.setWasOpened(true);
				}
				add(window, 0);
				try { 
					window.setClosed(false); 
				}
				catch (Exception e) { logger.log(Level.SEVERE,e.toString()); }
			}
			window.show();
			// bring to front if it overlaps with other windows
			try {
				window.setSelected(true);
			} catch (PropertyVetoException e) {
				// ignore if setSelected is vetoed	
			}
		}
	}