2015-01-22 Implemented some elements of robotic help in MSP. Created Robot.java, RobotConfig.java, robots.xml, TabPanelRobotic.java, RobotWindow.java, RobotDisplayInfoBean.java, RobotMapLayer.java, RobotTableModel.java, etc..
Note 1: In this implementation, Robot class extends Unit and reuses a lot of codes used by Person class. Therefore, I incorporated the use of the robot object into classes that the person object will use. 

Note 2: createInitialRobots() is NOT enabled in UnitManager.java as I'm still in the stage of debugging. The simulation is unaffected by the changes in this build.