2015-01-26a Embedding some JavaFX objects in MSP by creating a testing JFrame in MainWindowFX. The simulation runs normally and seems unaffected.
A. Changed to MarsProject.java to load up MainWindowFX.java.  

B. Loading MarsProject on EDT ahead of time in Main() so that when MainWindowFX object is created, it's already on EDT.  

C. Currently, JFXPanel is NOT created until calling initAndShowGUI() in MainWindowFX.java. We can declare JFXPanel and its JavaFX stuff inside the MarsProject.java if we want.