Refactoring construction so that stages can be created without all of the required construction materials, with the materials being added into the construction site as they become available.
Reducing demand for construction material goods based on potential for construction stages.
Adding new demand for construction material goods when they are required for an existing construction site.

Not allowing construction or salvage missions to begin until after ten sols from the start of the simulation to allow for good values to propagate fully.

Fixing bug in RandomUtil.getWeightedRandomItem when negative weight values are in the given item-weight map.

Fixing bug in ObserveAstronomicalObjects that was failing to return a valid observatory building.