2015-02-17a Added Medicbot. Dedicated for each bot type to do only related job tasks.
Attempted to isolate unrelated robot tasks by setting task and mission penalties for RobotJob.java to 0D. 

Bots should not have a job task that's NOT designed/programmed to do--unless the bot is re-programmed (a feature not coded in yet) by an engineer.

Assigned PrepareDessert jobtask to Chefbot.

Added Medicbot and assigned PrescribeMedication jobTask to Medicbot.

Fixed error in loading and assigning robotType from RobotConfig in createConfiguredRobots() in UnitManager.