2015-02-24 (A). Added toLowerCase() in loadAmountResources() in AmountResourceConfig.java in case if there are problems lurking somewhere for matching an amount resource name having uppercase letters with lowercase letters. This is in response to a user's ticket #54 Mineral Processing Bug. It appears that this smelting bug no longer exists as of now. This bug affected the release version 3.07 build 3118.
(B). Created getRobotMetaTasks() and initializeRobotMetaTasks() in MetaTaskUtil.java for bot's task selection in TaskManager.java. 

(C). Rewrote calculateProbability() in TaskManager.java for robots to call up getRobotMetaTasks() instead of getMetaTasks().

(D). Used if (robot.getBotMind().getRobotJob() instanceof ___bot) to limit bots to perform specified task in CookMealMeta, PrepareDessertMeta, ProduceFoodMeta, TendGreenhouseMeta.

(E). Initialized robot ahead of person in constructInitialUnits() in UnitManager.

(F). Added ArrivingSettlementEditingPanel.numOfRobots = # of Bots in messages.properties. 

(G). Reverted MarsProjectStarter back to loading instead of MarsProject instead of MarsProjectFX. 

(H.) Deleted Job ComboBox display and selection for bots in TabPanelActivity.java.

(I.) Reverted the memUsed calculation back to what it used to be in MainWindow.java. It is considered "normal" for the value of memUsed to rise and fall sharply within a second (possibly due to the peculiar activity of Java VM's garbage collection), as it can be inspected in detail using VisualVM. 

(J.) Disabled the use of splashscreen if loading the sim via MarsProjectFX.java.