2015-03-15b Resolved desserts loading issues and minor fixes.
(A). Dessert-related changes

(1). Added codes to compute the total amount of desserts when in  hasEnoughSupplies() and getRemainingSettlementAmount() in LoadVehicleGarage.java

(2). Added codes to skip checking for desserts in hasEnoughResources() in VehicleMission.java.

(3). Note: still fixing the display of correct amount of food and desserts in the Monitor Tool's "All Vehicle" Tab.

(B). Increased DOWN from 0.005 to 0.007 in Cooking.java and PreparingDessert.java to lessen the probability of expired meal/desserts.

(C). Reduced the amount of dessert cargo capacity on various vehicles in vehicles.xml.

(D). Commented out disposeMarqueeBanner() in MainDesktopPane.java.