2015-03-18a Added day night transition effect to Settlement Map Tool as an option.
(1). Created DayNightMapLayer class and added codes in SettlementMapPanel.java and SettlementTransparentPanel.java and SettlementWindow.java. 

(2). Re-used the same algorithm employed by ShadingMapLayer.java for the Navigator Tool.

(3). Added option to toggle on and off DayNightMapLayer in the label combobox on Settlement Map Tool 

(4). Temporarily commented out setMaximizable(true) in SettlementWindow.java. Enabling it seemed to interfere with the ShadingMapLayer mechanism.


Note 1 : needs further optimization or uses alternate algorithm to bring down the heavey CPU Utilization when the DayNightMapLayer was turned on.

Note 2 : the shadow cast needs to be less spherical.