2015-03-24 Patched issues with bots inside a vehicle.
(1). Added BOTS column and added getRobotCrewNum() in getValueAt() in Vehicle Tab on Monitor Tool in VehicleTableModel.java.

(2). Added the usage of getRobotCrewNum() for bots in a number of vehicle related classes.

(3). Cloned TabPanelBots.java from TabPanelCrew.java.

(4). Switched to loading TabPanelBots instead of TabPanelCrew in VehicleWindow.java's constructor if bots are using the vehicle.

(5). NOTE: for now, our codes do NOT allow mixing the bots and people together in one vehicle. Bots can only be with bots in a vehicles.