2015-04-11 Added in more validation codes in ScenarioConfigEditorFX and SimulationConfigEditor, paving the way for saving settlement data in Host Server's settlement central registry in the next few builds.
1. Set the ground rule to allow a new settlement be built at a minimum of 0.1 degree (or 5.7 km apart) from an existing settlement. Currently, the Editor allows only decimal inputs in latitude/longitude but at the start of simulation, it truncates all decimal places and/or ignore the user's input and replaces with internal default location for new settlements. 

2. Accounted for the following error conditions and assisted users in selecting latitude and longitude:

a). checking for repeating latitude and longitude already occupied by another settlement(s) on the table. 

b). checking if 0 S and 0 N (and 0 E and 0 W) are being selected in the table since they are identical Latitude location.

c). checking if the second character from the last is a digit or a letter

d). checking if the last character is a digit or a letter

e). checking if a garbage mixed string of letter and digit is input

If found, call setError() and disable "Create New Settlement" button temporarily until a new correct latitude is entered.

f). checking if the decimal place is > 2. If true, correct it to 1 decimal place right away

3. Modified ui elements in css/fxml files.