2015-04-14 Fixed issues with multiplayer mode.
(1). Modified ScenarioConfigEditor for multiplayer mode.

a. Call loadDefaultSettlements() if the user is the first client, or else call addExistingSettlement().

b. Added loading of existing settlements (from other client machine) in ScenarioConfigEditor in multiplayer mode;

c. Added loadExistingSettlements() and setExistingSettlements()

d. Changed the function of the defaultButton (and changed name to refreshDefaultButton) into dual use: for (1) loading the default two settlement settlements. (2). refreshing the global map for any existing settlement.

e. Added color to JTable and JHeader and grid lines

(2). Added the requirement of setting up a playerName when a client is attempting to log in the host server. 

a. Created idMap to store player id and usersname in MultiplayerServer:
Map<Integer, String> idMap = new HashMap<>();

b. after successfully loading MultiplayerClient, it will automatically go back to Main Menu and run "New Simulation" and fire up ScenarioConfigEditor.

c. Modified getNewID() to assign each client a player/client ID as soon as he is connected to the host server. If the connection drops, the player ID will be returned to the server.

d. Enabled the client to log into a host server at a different IP other than the localhost (if it's allowed through firewall).


(3). Currently, the client can send the following three commands to the host server :

a. get : obtain a list of all existing settlements
b. register (playerName) : register with the server and be assigned with a client id
c. new (new settlement info) : update the server with a newly created settlements
d. bye (playerName) : disconnect from the server (when quitting the client)
e. s : obtain the number of existing settlements.

(4). currently, Neither client id nor playerName is NOT unique to each client machine. client id is linked to a playerName and is unique to each player. A client machine can house more than one player but would have the same client IP address. In future this will change.

(5). Added new networking package folder in mars-sim-core and moved two classes there.