2015-04-23 Modified validateLandingLocation() to round latitude and longitude to one decimal and performed checking of other client's existing settlement coordinates in ArrivingSettlementEditingPanel.java in order to make settlements' coordinates unique across the multiplayer simulation.
Created a separate static ClientRegistry class keep track of the settlement list. Can be called by MultiplayerClient and other class.

Fixed the display of the number of bots and its textfield in ArrivingSettlementEditingPanel.java.

Added getLatitudeDouble() and getLongitudeDouble() in Coordinates.java. Called by MultiplayerClient.java.

Added calling ColumnResizer.java in ScenarioCofigEditorFX.java to automatically adjust the wide of the JTable.