Fixing problem with caloric energy being expended much faster than gained from eating.
Making adding and subtracting caloric energy for a person more linear rather than exponential.  We may want to make it exponential if it's more accurate to model it that way, but we need to test the equations to make sure adding and subtracting energy is symetrical.

Changing starvation to be triggered only from hunger rather than hunger or caloric energy.  This was causing frequent brief moments of starvation, where it should happen only after a long period without eating.  We might want to reconsider this later if we make changes to caloric energy levels.

Changing the Cooking.cleanup method to be called when meal time is over rather than from the CookingMeal task.

Fixing problems with the CookingMealMeta.getProbability not returning a positive result.

Adding check in CookingMealMeta.getProbability for if there are enough ingredients to cook a meal.

Adding check in CookingMealMeta.getProbability for if there are enough meals already cooked at the settlement.

Refactoring EatingMealMeta.getProbability based on hunger and caloric energy levels.

Adding check in Cooking building function for total number of cooked meals at settlement to compare against max servings.

Fixing bug with bad or expired cooked meals not being properly removed from the cookedMeals list.