2015-04-28 (1). Added shaking animation to the login dialog if a user inputs wrong playername/password pair in loginDialog() in MultiplayerClient.java. Added ShakeTransition class.
(2). Added alert dialog asking for final confirmation before quitting msp if a user clicks the top right hand quit icon in the login dialog and the client connection dialog.

(3). Added a new class BuildingWindow.java and used UnitWindowFactory.java to create an instance of it.

(4). Added timeTagMap in MultiplayerServer.java to track how long a player was last connected with the server.