2015-05-08 Multithreading refactoring
(1). Since the initial adoption of Java 8 after the last release, a majority of classes such as Simulation.java and MarsProjectFX, MarsClock, etc. used to be running inside JavaFX Application Thread. 

(2). This refactoring serves to relocate those classes (as well some background task codes in mars-sim-ui) to run outside of the JavaFX Application Thread, thus paving the way for more responsive UI.

(3). Used ThreadPoolExecutors in ScenarioConfigEditorFX.java and in Simulation.java to create concurrent threads so as to offload work from JavaFX Application Thread.