2015-05-09 More rework on settler's career and job reassignment.
(1). Added codes in JobAssignment.java to record the following three additional info: 

a. the status of the new job reassignment (either "Pending" or "Approved"). 
b. the name of the person who authorizes the reassignment. 
c. the date approved. 

(2). Modified codes in TabPanelCareer to handle the submission of job reassignment. 

(3). Added a one-day delay between the submission between the time the job reassignment is submitted and the enactment of the change of job type. 

(4). Replaced JLabel with JTextfield for better look in General tab and Career tab in person unit window. 

(5). Will soon add a new ai task called "ReviewJobAssignment.java" for the commander/mayor or the chief of the division/cabinet to approve these job reassignment.