Fixing endless loop bug in Role.java that was causing a StackOverflowException.  Role.setNewRoleType will only set a new role type when it is different from the person's current role type.
Fixing a couple NullPointerException bugs in TabPanelCareer.java:
Changing person.getSettlement() to person.getAssociatedSettlement(), which should return the person's home settlement in almost all cases.
Changing status.equals("Accepted") to status == "Accepted", which should work even if the status string is null.

Changing the degree symbol in TemperatureGauge.java to a unicode degree symbol as the given one isn't valid UTF-8, which is required for maven on the command line.