Adding "Construction" and "Operations" to the favorite activity list.  "Operations" is for general settlement work, such as loading and unloading vehicles.  "Construction" is for construction and salvaging buildings.
Refactoring favorite modifiers through task meta classes getProbability methods.  Tasks that are associated with a favorite activity will have their overall probability doubled if that is the person's favorite activity.  This replaces the +50 probabilties and such.

Modifying the SleepMeta.getProbability method to try to make it more likely that a person will sleep when it's dark out, and therefore sleeping patterns will start matching up with the day/night patterns of the person's location.  Also making an exception for astronomers who will sleep during the daytime and stay awake at night.