2015-05-20a Improved weather modeling.
(1). Revised optical depth calculation, added getOpticalDepth(), opticalDepthMap getPreviousSolarIrradiance(), saveSolarIrradiance() in SurfaceFeatures.java.

(2). Added the display of the value of optical depth (aka visibility) and solar irradiance in TabPanelWeather.java.

(3). Remodeled the effect of Diffuse Solar Irradiance G_dh in getSolarIrradiance() in SurfaceFeatures.java so that 
there is more average sunlight in each sol.

(4). Renamed radius to instantaneousSunMarsDistance in in OrbitInfo.java.

(5). Fixed an error in calculating instantaneous radius (formerly called radius) of Mars (Mars-Sun distance). This error caused the calculated solar irradiance on Mars to exceed the beam solar irradiance at the top of the Mars atmosphere. 

(6). Added INSTANTANEOUS_RADIUS_NUMERATOR and DEGREE_TO_RADIAN in OrbitInfo.java.

(7). Remodeled part 1 of the temperature calculation in Weather.java to make it dependent upon the amount of solar irradiance, replacing the original equation which is purely based on the time of day and longitude (the equation  seems to have error).