Fixing NullPointerException in Weather class when creating a new simulation from the UI.  This involved changing temperatureCacheMap and airPressureCacheMap variables from static to transient (not being saved in the saved file object stream) and using lazy instantiation so they are only instantiated when first used in a new simulation.
Increasing starting probability for the following classes: LoadVehicleGarage, LoadVehicleEVA, UnloadVehicleGarage, UnloadVehicleEVA, ConstructBuilding, and SalvageBuilding.  This means that when there is a mission rover needing loading or unloading, or an active construction/salvage mission, people will be more likely to assist when not on that mission.

Removing a couple of characters that were unmappable to UTF8 from MeteoriteImpactImpl.java and Building.java.  These were preventing Maven compilation from the command line on my Linux laptop.