2015-06-09 Added codes to Radiation Exposure Modeling and fixed a bug in Meteor Impact Modeling. Added a row header column with numbers for all JTables in Monitor Tool.
(1). Added new classes RadationEvent.java, BodyRegionType.java. 

(2). Modified column resizing in TableTab.java to set a minimum width of 100 so that the column width of "Task" can show more text.

(3). Used adjustColumnPreferredWidths() within setTableStyle() in TableTab.java.

(4). Added column sorting capability to all JTable found in person unit window and settlement unit window.

(5). Added calling MultisortTableHeaderCellRenderer.java in any TabPanel___.java class with JTable. 

(6). Added codes in addDose() and timePassing() in RadiationExposure.java. Set checking for exposure once a day at between 100 and 110 millisols.

(7). Corrected how to set and reset a boolean that enables a meteorite impact to occur in checkForMeteoriteImpact() in Building.java.

(8). Renamed "meteorite impact" to "meteorite impact damage" in malfunctions.xml and in getMeteoriteImpactMalfunction() in Building.java.

NOTES: 

A.  still need helps in finding equations of the probability distribution of different sizes of the meteorites

B.  still need to build an incredible model in relating how the size of the meteorites and its speed of impact would damage a building.