2015-06-09a Fixed bug in computing L_s and season. Misc fixes.
(1). Fixed bug in computing and updating instantaneousSunMarsDistance in addTime() in OrbitInfo.java.

(2). Commented out blocks of code below in numerous methods in Weather.java since the value of millisols (as global variable) is already being recomputed and updated in timePassing().

(3). Refactored getSeason() in MarsClock.java based on the value of L_s. L_s was initially set to be 252.58 deg (which corresponds to Earth date 2043-Sep-30) at the beginning of the sim. 

Note: At L_s = 252, the Northern Hemi season should be “Late Autumn” at the beginning of the sim.

(4). Implemented Serializable for RadiationEvent.java.