2015-06-09b  Fixed temperature fluctuation issue with lander hab and EVASuit. Added a row header column with number in Monitor Tool.
(1). Reworked the codes in timePassing() in Heating.java to make sure it wiould skip calling adjustThermalControl(). 

(2). Fixed bad temperature warning by replacing the calling of weather.getTemperature with weather.calculatedTemperature() in getTemperature() in EVASuit.java at the start of the sim.

(3). Used adjustColumnPreferredWidths() within setTableStyle() in TableTab.java.

(4). Commented out column sorting capability in TabPanelThermalSystem ()and previously TabPanelPower) until the exception is resolved.

(5). Modified column resizing in TableTab.java to set a minimum width of 100 so that the column width of "Task" can show more text.