2015-06-18 Misc fixes. Fixed weather and solar irradiance issues. Fixed ConcurrentModificationException in MasterClock.
(1). Changed the name of one assignJob() method to checkJob() in Mind.java.

(2). Added getInitialJob() in Mind.java and called two times by UnitManager.java at the start of the sim.

(3). Set an upper limit of the value of solar irradiance at 586 W / m2 in getSolarIrradiance() in SurfaceFeatures.java.

(4). Set an lower limit of the value of solar irradiance at 20.94 W / m2 in getSolarIrradiance() in SurfaceFeatures.java when Mars is at the horizon as it enters the twilight zone. 

(5). Tweaked the fluctuation of the value of tau (optical depth).

(6). Normalized sunlight to 586 (given by SurfaceFeatures.MEAN_SOLAR_IRRADIANCE) in GlobalDisplay.java and DayNightMaplayer.java. Renabled getSolarIrradiance() in these two classes for testing and comparison with getSurfaceSunlight(). 

(7). Added more constants in Heating.java for future modeling of the thermal aspect of the inflatable transparent greenhouse rooftop.

(8). Added windSpeedCacheMap and windDirCacheMap and called them in computeWindSpeed(), computeWindDir() in Weather.java so that the wind speed/dir at different locations are decoupled.

(9). Added wind speed as a contributing factor in calculateTemperature() in Weather.java.

(10). Added randomness to the temperature in the dark polar region.

(11). Converted HashMap to ConcurrentHashMap in Weather.java.

(12). Fixed ConcurrentModificationException by replacing ArrayList with CopyOnWriteArrayList in MasterClock.java and Weather.java.