2015-06-25 Fixed loading and saving issues due to recent multithreading changes.
(1). Resolved default sim loading issue. Added getFileSize() in Simulation.java to obtain the size of the default.sim file when selecting option 2 in Main Menu to load the default saved sim. Calculated the time delay needed to repopulate the saved data before loading the various Tools in MainScene. 

(2). Fixed the missing file extension issue in javaFX mode when using "save as" to save an unique filename. Removed the ExtensionFilter of all files (*.*) in saveSimulationProcess() in MainScene.java so that the default extension .sim is always chosen.

(3). Removed the param in autosaveSimulation() in MasterClock.java since the param is null anyway.