2015-06-26a Implemented variable sleep time in MasterClock by using the sleep time codes in this book at http://fivedots.coe.psu.ac.th/~ad/jg/code/index.html Ch2 P.33-34
(1). Used System.nanoTime() twice within the run() in ClockThreadTask in MasterClock.java to gauge the time difference between the beginning and the end of a frame. 

(2). The time difference will be used for adjusting the sleep time at the end of a frame so that the duration of each frame (including the sleep time) will stay constant.

(3). The default sleep time is set to 25 ms or 40 Hz for now.

(4). TODO: for a simulation with, say, 300 people, one should tune the default sleep time to below 10 ms for more responsive UI. Need more testing to find the relationship between sleep time and # of population for an optimal UI.