2015-06-30a Revised behavior in TabPanelSchedule.java.
(1). Set realTimeUpdateCheckBox to true (checked) by default.

(2). Binded the "true" state of realTimeUpdate to selecting the sol combobox as "today" (i.e. the current sol). 

Usage Notes:
a. If the user checks the realTimeUpdateCheckBox, realTimeUpdate is set to true and today's schedule is automatically selected and vice versa.

b. while the user is viewing the current sol's schedule and the martian time passes the midnight, the sol selection will not change (it will NOT advance the next sol's schedule) but the sol combobox will get incremented. Also the realTimeUpdateCheckBox will be automatically unchecked.

(3). Added more conditional checking for the frequencyMap in setTaskStatus() in Preference.java.