Varying the refresh rate of the map panel based on if it is displayed in the Mars Navigator tool, the Mission tool, or the Mission creation wizard.  This now allows for a high refresh rate for smooth navpoint placement when creating a new rover mission.
Setting construction vehicles to display in Vehicle field in the Mission tool when a construction or salvage mission is selected.

Setting construction vehicles to display in the Vehicle column of the Mission tab of the Monitor tool for a construction or salvage mission.

Fixing NullPointerException in SurfaceFeatures.getSurfaceSunlight when the simulation has not been fully constructed but the UI is calling the method.