2015-11-04 Fixed IllegalStateException "ConstructionBot ___ is not currently in a building". Added junit in mars-sim-main's pom.xml.
1. Provided a workaround in getClosestWalkableAvailableAirlock() in Settlement.java.

2. Reworked codes in getBuilding(robot) in BuildingManager.java. 

---------------------------------
Notes: 
---------------------------------
A. The original concept was that a robot is thought to need RoboticStation function to "survive", much like a person needs LifeSupport function.

B. We might have since deviated from this rule. Besides, a ConstructionBot has to be outside in a construction site where it's exposed to the martian air. It shouldn't be an issue at all if the bot is designed to aid the settler's in construction projects.

C. In case a bot is in a settlement or in a building where there is no RoboticStation function, the codes in getBulding() will still find the bot.

----------------------------------

3. Minor tweaking on UI look in PopUpUnitMenu.java, UnitInforPanel.java, BuildingPanel.java.

4. Added titleLabel in TabPanelBots.java and TabPanelCrew.java.

5. Centered balloon tooltip in BuildingPanelFarming.java. 

6. Darkened the green color in CenterMap.png

7. Added junit in mars-sim-main's pom.xml to avoid error for MacOSX

<dependency>
	<groupId>junit</groupId>
	<artifactId>junit</artifactId>
	<version>4.12</version>
</dependency>