2015-12-17a Beefed up the muscle of the chat box to recognize settlement names.
Revised hasRobotName() to match the input robot name and added checkRobotName() in Settlement.java.

Revised parseText() to recognize settlement's names and backslash symbol (indicative of a name) in ChatBox.java.

Added a Broadcast button to the left of the textfield in the chat box as an alternative for submitting an inquiry in ChatBox.java.