2015-12-18 Overrode the use of the Escape key (ESC) for  pausing/unpausing the sim. Misc Fixes. Switched to making only one instance of Stage in across the MarsProjectFX, MainMenu and MainScene.
------------------------------------------
Concept Notes:
------------------------------------------

A. the default use of ESC to exit the full screen mode in JavaFX has already been supplanted by MainSceneMenu in that Ctrl+F (or Command+F in Mac) is used for toggling the full screen mode.

B. The Stage that comes from the JavaFX Application Thread in MarsProjectFX will be used in MainMenu to host the main menu's Scene instance. When users start a sim, this same stage will be used again to host the main scene's Scene instance. 

C. We want to prevent user from inadvertently choosing any menu options to start another sim while a sim is stilll on-going.

D. MainMenu class is at mars-sim-main and MainScene class is at mars-sim-ui. The Maven allows only one way dependency. In our case, we have mars-sim-ui depending on mars-sim-main. As a result, once we have loaded MainScene.java. Programmatically, one cannot go back and access an object/method in MainMenu.java.

E. This screen switching is arguably a cleaner way of ensuring that users will not access the main menu. 

F. This scene/stage configuration is not set in stone and may be altered to meet the future needs. 

------------------------------------------
IMPORTANT NOTICE:
------------------------------------------
1. Still tracking down the culprit of a mysterious/unknown "program halt" that has been plaguing the last few builds. 
------------------------------------------

----------------------------------------
Code Changes:
----------------------------------------

1. Added setFullScreenExitKeyCombination() and setFullScreenExitHint() in initAndShowGUI() in MainMenu.java to override the use of ESC for pausing/unpausing on the stage of the MainScene.

2. Set the font size to 20 in mainmenu.css if the mouse is not hovering over any of the main menu's options.

3. Added checking for previous pause state in saveSimulationProcess() (used by both regular save and autosave process) in MainScene.java and saveSimulation() in Simulation.java. This effort may be redundant/unhelpfully in trying to debug the aforementioned unknown program halt.

4. Added validate() and repaint() in openAnnouncementWindow() to avoid ugly black border to form around the annoucement window.

5. Removed the need of declaring a new stage instance just for the main menu's scene. Reworked MainMenu's constructor and a couple of methods to reuse the same Stage instance. Relocated a few statements on the stage setup from MainMenu to MainScene.