2015-12-22 Added the 3 sols and 10 sols average deltas (for each the 9 critical resources) in Marquee Ticker at the bottom of the Settlement Map Tool.
1. Modified getOneResource() and createResourceTicker() to accommodate the 3 and 10 sols delta points in MarqueeTicker.java.

2. Modified getOneResource() and createResourceTicker() in MarqueeTicker.java to display, i.e. "--- (1 sol)", in grey color if the data is not available yet.

3. Added 20 whitespaces right before the O2 resource label so that mars-simmers won't miss out spotting the first resource in getOneResource() in MarqueeTicker.java.

4. Modified getAverage() to compute the 3 and 10 sols average.

5. Set ROBOT_LABEL_COLOR to a brighter orange color in LabelMapLayer.java so that robot dots are distinguishable from the settlement map's background color.

------------------------------------------------------------
CONCEPT NOTES: 
a. The delta for 3 sols is the difference of between the average mass of a resource 3 sols ago and today's average

b. This way, mars-simmers can project the surplus or shortfall in longer terms.

c. The 1 sol delta will be shown as "--- (1 sol)" in grey color until sol 2. 

d. The 3 sols delta will be shown as "--- (3 sols)" in grey color until sol 4. 

e. The 10 sols delta will be shown as "--- (10 sols)" in grey color until sol 11. 

------------------------------------------------------------