2016-01-16. Implemented a dashboard for each settlement.
---------------------------------------------------
Concepts: 
---------------------------------------------------
i. Dashboard is like the front page of a newspaper for displaying highlighted information regarding the settlement.

ii. The "Overall Objective" of the settlment is a new concept to be further developed. For now, it is added to the Dashboard as a combobox so that it can be changed by mars-simmers.

iii. For now, there are five possible objectives for a settlement:
a. Crop Farm
b. Manufacturing Depot
c. Research Campus
d. Transportation Hub
e. Trade Center

iv. A settlement's objective will affect the strategies of how a settlement will be evolved over time. 

v. Settlement strategies will be implemented later. They should be tunable by mars-simmers.

vi. Currently, Dashboard tab is only visible in JavaFX mode.
---------------------------------------------------

1. Added enum class ObjectiveType and Objective interface class.

2. Made use of JFXPanel for the first time within a tab inside the settlement unit window. 

3. Added TabPanelDashboard.java and have it open at the first tab when loading settlement unit window. 

4. Employed JavaFX UI coding instead of Swing UI in TabPanelDashboard.java.

5. Added loading TabPanelDashboard in SettlementUnitWindow.java in case of JavaFX mode.